chore: move cache-memory.test.ts to unit tests
