--- packages/astro/test/fixtures/cache-memory/astro.config.mjs
@@ -1,10 +1,0 @@
-// @ts-check
-import { defineConfig, memoryCache } from 'astro/config';
-
-export default defineConfig({
-	experimental: {
-		cache: {
-			provider: memoryCache(),
-		},
-	},
-});
