--- packages/astro/test/fixtures/cache-memory/package.json
@@ -1,8 +1,0 @@
-{
-	"name": "@test/cache-memory",
-	"version": "0.0.0",
-	"private": true,
-	"dependencies": {
-		"astro": "workspace:*"
-	}
-}
