--- packages/astro/test/fixtures/cache-memory/src/pages/head-cached.ts
@@ -1,6 +1,0 @@
-export const prerender = false;
-
-export const GET = async (context) => {
-	context.cache.set({ maxAge: 300, tags: ['head-cached'] });
-	return Response.json({ timestamp: Date.now() });
-};
