--- packages/astro/test/fixtures/cache-memory/src/pages/invalidate-path.ts
@@ -1,6 +1,0 @@
-export const prerender = false;
-
-export const POST = async (context) => {
-	await context.cache.invalidate({ path: '/cached' });
-	return Response.json({ invalidated: true });
-};
