--- packages/astro/test/fixtures/cache-memory/src/pages/no-cache.ts
@@ -1,6 +1,0 @@
-export const prerender = false;
-
-export const GET = async (context) => {
-	// No cache.set() — should pass through without caching
-	return Response.json({ timestamp: Date.now() });
-};
