--- packages/astro/test/fixtures/cache-memory/src/pages/opt-out.ts
@@ -1,6 +1,0 @@
-export const prerender = false;
-
-export const GET = async (context) => {
-	context.cache.set(false);
-	return Response.json({ timestamp: Date.now() });
-};
