--- packages/astro/test/fixtures/cache-memory/src/pages/with-cookie.ts
@@ -1,8 +1,0 @@
-export const prerender = false;
-
-export const GET = async (context) => {
-	context.cache.set({ maxAge: 300, tags: ['cookie'] });
-	const response = Response.json({ timestamp: Date.now(), nonce: Math.random() });
-	response.headers.set('Set-Cookie', 'session=test; Path=/; HttpOnly');
-	return response;
-};
